+++ code/pom.xml
@@ -60,6 +60,17 @@
 			<artifactId>spring-security-core</artifactId>
 			<version>5.5.2</version>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.security</groupId>
+			<artifactId>spring-security-config</artifactId>
+			<version>5.5.2</version>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.security</groupId>
+			<artifactId>spring-security-test</artifactId>
+			<version>5.5.2</version>
+		</dependency>
+		
 
 		<dependency>
 			<groupId>org.jsoup</groupId>
