+++ code/src/main/webapp/WEB-INF/spring/security-context.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<<<<<<< Updated upstream
 <beans xmlns="http://www.springframework.org/schema/beans"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xmlns:security="http://www.springframework.org/schema/security"
@@ -9,15 +10,15 @@
 
     <bean id="customAccessDenied"
           class="com.company.security.CustomAccessDeniedHandler"/>
+=======
+<beans xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xmlns="http://www.springframework.org/schema/beans"
+	xmlns:security="http://www.springframework.org/schema/security"
+	xsi:schemaLocation="http://www.springframework.org/schema/security http://www.springframework.org/schema/security/spring-security.xsd
+		http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
+>>>>>>> Stashed changes
 
     <security:http>
-        <security:intercept-url pattern="sample/all"
-                                access="permitAll"/>
-        <security:intercept-url pattern="sample/user"
-                                access="hasRole('ROLE_USER')"/>
-        <security:intercept-url pattern="sample/admin"
-                                access="hasRole('ROLE_ADMIN')"/>
-        <security:access-denied-handler ref="customAccessDenied"/>
         <security:form-login authentication-failure-url="/login?error=true"/>
         <security:csrf/>
     </security:http>
